DBMS_SQL/requirements.txt
fastapi
uvicorn[standard]
streamlit
psycopg[binary]
python-dotenv
pandas
requests
